pyproject.toml
[project]
name = "pyutils"
version = "0.1.0"
description = "Python utility scripts"
requires-python = ">=3.13"
dependencies = [
    "boto>=2.49.0",
    "boto3>=1.42.3",
    "jinja2>=3.1.6",
    "nested-dict>=1.61",
    "nose>=1.3.7",
    "pandas>=2.3.3",
    "psutil>=7.1.3",
    "pytest>=9.0.1",
    "pytest-xdist>=3.8.0",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "simplejson>=3.20.2",
    "six>=1.17.0",
    "slacker>=0.14.0",
    "urllib3>=2.5.0",
]

[tool.poetry.dependencies]
python = "^3.9"
Jinja2 = "^3.1.1"
PyYAML = "^6.0"
boto = "^2.49.0"
boto3 = "^1.21.31"
nested_dict = "^1.61"
nose = "^1.3.7"
psutil = "^5.9.0"
pytest = "^8.1.1"
python-dateutil = "^2.8.2"
requests = "^2.27.1"
simplejson = "^3.17.6"
six = "^1.16.0"
slacker = "^0.14.0"
urllib3 = "^1.26.9"
black = "^24.3.0"
flake8 = "^7.0.0"
pylint = "^3.0.1"
pandas = "^2.0.1"
rich = "<=14.0.0"
pytest-xdist = "^3.3.1"

[tool.ruff]
lint.select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # Pylint
    "PL"
]
lint.ignore = [
    # tests use assert
    "S101"
]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
]
